# Perf improvements for small or value-type POCOs