--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterOfT.WriteCore.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterOfT.WriteCore.cs
@@ -24,7 +24,7 @@ namespace System.Text.Json.Serialization
 
         internal bool WriteCore(
             Utf8JsonWriter writer,
-            T value,
+            in T value,
             JsonSerializerOptions options,
             ref WriteStack state)
         {