… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonClassInfo.Cache.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonClassInfo.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfo.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoOfT.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandlePropertyName.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.ByteArray.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.Helpers.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.Stream.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.String.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.Utf8JsonWriter.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializerOptions.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMemberAccessor.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReflectionMemberAccessor.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/WriteStack.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Literal.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.String.cs
… src/libraries/System.Text.Json/tests/Serialization/CacheTests.cs